Extract service for finding all template file paths
